Migrate the bot to the new API.

+++ src/main/java/ru/javazen/telegram/bot/repository/SubscriptionRepository.java
@@ -9,5 +9,5 @@ import java.util.List;
 public interface SubscriptionRepository extends CrudRepository<Subscription, MessagePK> {
     List<Subscription> findAllBySubscriptionPK_ChatId(Long chatId);
 
-    Long countAllBySubscriptionPK_ChatIdAndUserId(Long chatId, Integer userId);
+    Long countAllBySubscriptionPK_ChatIdAndUserId(Long chatId, Long userId);
 }
