+++ src/main/java/ru/javazen/telegram/bot/service/SubscriptionService.java
@@ -8,7 +8,7 @@ import java.util.List;
 public interface SubscriptionService {
     void createSubscription(Subscription template);
 
-    List<Subscription> catchSubscriptions(Long chatId, Integer userId, String text);
+    List<Subscription> catchSubscriptions(Long chatId, Long userId, String text);
 
     void saveSubscriptionReply(MessagePK subscriptionPK, int replyMessageId);
